+++ src/app/page.tsx
@@ -2,6 +2,7 @@
 
 import type React from "react"
 import { useState } from "react"
+import { useRouter } from "next/navigation"
 import { Eye, EyeOff, Coffee, User, Mail, Lock, ArrowLeft } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
@@ -13,11 +14,89 @@ export default function AuthForm() {
   const [showPassword, setShowPassword] = useState(false)
   const [showConfirmPassword, setShowConfirmPassword] = useState(false)
   const [activeTab, setActiveTab] = useState("login")
+  const [alert, setAlert] = useState<{
+    type: "success" | "error"
+    message: string
+  } | null>(null)
+  const router = useRouter()
 
-  const handleSubmit = (e: React.FormEvent, type: string) => {
+  const handleSubmit = async (e: React.FormEvent, type: string) => {
     e.preventDefault()
-    console.log(`${type} form submitted`)
-    // Handle form submission logic here
+    setAlert(null)
+
+    const form = e.target as HTMLFormElement
+
+    if (type === "login") {
+      const email = (form.elements.namedItem("login-email") as HTMLInputElement).value
+      const password = (form.elements.namedItem("login-password") as HTMLInputElement).value
+
+      const res = await fetch("/api/auth", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ type: "login", email, password }),
+      })
+      const data = await res.json()
+      if (res.ok) {
+        router.push("/pages/inventory")
+      } else {
+        setAlert({
+          type: "error",
+          message: data.error || "Login Faild"
+        })
+      }
+    } else if (type === "register") {
+      const name = (form.elements.namedItem("register-name") as HTMLInputElement).value
+      const email = (form.elements.namedItem("register-email") as HTMLInputElement).value
+      const password = (form.elements.namedItem("register-password") as HTMLInputElement).value
+      const confirm = (form.elements.namedItem("confirm-password") as HTMLInputElement).value
+
+      if (password !== confirm) {
+        setAlert({
+          type: "error",
+          message: "Passwords do not match"
+        })
+        return
+      }
+
+      const res = await fetch("/api/auth", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ type: "register", name, email, password }),
+      })
+      const data = await res.json()
+      if (res.ok) {
+        setAlert({
+          type: "success",
+          message: data.success || "Register Sucessfull"
+        })
+        // setActiveTab("login")
+      } else {
+        setAlert({
+          type: "error",
+          message: data.error || "registration failed"
+        })
+      }
+    } else if (type === "reset") {
+      // Implement reset logic if needed
+      const email = (form.elements.namedItem("reset-email") as HTMLInputElement).value
+      const res = await fetch("/api/forgot-password", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ email }),
+      })
+      if(res.ok){
+        setAlert({
+          type: "success",
+          message: "Reset link sent to your email"
+        })
+      } else {
+        const data = await res.json()
+        setAlert({
+          type: "error",
+          message: data.error || "Failed to send reset link"
+        })
+      }
+    }
   }
 
   return (
@@ -59,7 +138,7 @@ export default function AuthForm() {
                   value="reset"
                   className="text-gray-300 data-[state=active]:bg-amber-600 data-[state=active]:text-white"
                 >
-                  Reset
+                  Forgot Password
                 </TabsTrigger>
               </TabsList>
 
@@ -118,6 +197,9 @@ export default function AuthForm() {
                       Forgot password?
                     </Button>
                   </div>
+                  {alert && (
+                    <div className="text-red-500 text-sm text-center">{alert.message}</div>
+                  )}
                   <Button type="submit" className="w-full bg-amber-600 hover:bg-amber-700 text-white">
                     Sign In
                   </Button>
@@ -217,6 +299,12 @@ export default function AuthForm() {
                       </Button>
                     </label>
                   </div>
+                  {alert && (
+                    <div className={`text-sm text-center ${alert.type === "success" ? "text-green-500" : "text-red-500"
+                      }`}
+                    >
+                      {alert.message}</div>
+                  )}
                   <Button type="submit" className="w-full bg-amber-600 hover:bg-amber-700 text-white">
                     Create Account
                   </Button>
